threads: split CThreadImplPosix into CThreadImplLinux
