+++ cmake/treedata/linux/subdirs.txt
@@ -7,6 +7,7 @@ xbmc/platform/linux/peripherals     platform/linux/peripherals
 xbmc/platform/linux/powermanagement platform/linux/powermanagement
 xbmc/platform/linux/sse4            platform/linux/sse4
 xbmc/platform/linux/storage         platform/linux/storage
+xbmc/platform/linux/threads         platform/linux/threads
 xbmc/platform/posix                 platform/posix
 xbmc/platform/posix/filesystem      platform/posix/filesystem
 xbmc/platform/posix/network         platform/posix/network
