+++ xbmc/platform/linux/threads/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(SOURCES ThreadImplLinux.cpp)
+set(HEADERS ThreadImplLinux.h)
+
+core_add_library(platform_linux_threads)
