+++ xbmc/platform/linux/threads/ThreadImplLinux.h
@@ -9,15 +9,21 @@
 #pragma once
 
 #include "threads/IThreadImpl.h"
+#include "threads/SingleLock.h"
 
-class CThreadImplPosix : public IThreadImpl
+class CThreadImplLinux : public IThreadImpl
 {
 public:
-  CThreadImplPosix(std::thread::native_handle_type handle);
+  CThreadImplLinux(std::thread::native_handle_type handle);
 
-  ~CThreadImplPosix() override = default;
+  ~CThreadImplLinux() override = default;
 
   void SetThreadInfo(const std::string& name) override;
 
   bool SetPriority(const ThreadPriority& priority) override;
+
+private:
+  CCriticalSection m_criticalSection;
+
+  pid_t m_threadID;
 };
